:recycle: EmailClient 의존성 격리
