+++ src/main/kotlin/com/wafflestudio/spring2025/common/email/client/JavaMailEmailClient.kt
@@ -1,8 +1,12 @@
 package com.wafflestudio.spring2025.common.email.client
 
+import com.wafflestudio.spring2025.common.email.exception.EmailErrorCode
+import com.wafflestudio.spring2025.common.email.exception.EmailServiceUnavailableException
+import jakarta.mail.MessagingException
 import jakarta.mail.internet.MimeMessage
 import org.slf4j.LoggerFactory
 import org.springframework.context.annotation.Primary
+import org.springframework.mail.MailException
 import org.springframework.mail.javamail.JavaMailSender
 import org.springframework.mail.javamail.MimeMessageHelper
 import org.springframework.stereotype.Component
@@ -21,26 +25,25 @@ class JavaMailEmailClient(
         fromEmail: String,
         fromName: String,
     ) {
-        logger.info("메일링 임시 중단으로 발송 생략: $to")
-        return
+        // 지금 AWS SDK를 쓰는 게 아니라 그냥 email config를 AWS SES의 SMTP 엔드포인트로 해놓은 상황
+        // (SDK가 아닌 SMTP 방식으로 SES 사용)
         try {
             val message: MimeMessage = javaMailSender.createMimeMessage()
-
-            val helper = MimeMessageHelper(message, true, "UTF-8")
+            val helper = MimeMessageHelper(message, "UTF-8")
 
             helper.setTo(to)
             helper.setSubject(subject)
-
             helper.setText(htmlContent, true)
-
             helper.setFrom("$fromName <$fromEmail>")
 
+            // AWS SES 샌드박스 모드에서는 발신/수신 주소 모두 사전 인증(Verified)이 필요합니다.
             javaMailSender.send(message)
-
-            logger.info("Gmail을 통해 메일 발송 성공: $to")
-        } catch (e: Exception) {
-            logger.error("Gmail 메일 발송 실패: $to", e)
-            throw RuntimeException("Email sending failed via Gmail", e)
+        } catch (e: MessagingException) {
+            logger.error("메일 구성 실패: to={}, subject={}", to, subject, e)
+            throw EmailServiceUnavailableException(EmailErrorCode.EMAIL_SERVICE_UNAVAILABLE)
+        } catch (e: MailException) {
+            logger.error("메일 전송 실패: to={}, subject={}", to, subject, e)
+            throw EmailServiceUnavailableException(EmailErrorCode.EMAIL_SERVICE_UNAVAILABLE)
         }
     }
 }
